% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{kenya}
\alias{kenya}
\title{Datasets for Use in Designing Surveys in Kenya}
\description{
Designing surveys require relevant datasets to be used as basis for
sample size calculations, sampling design, survey planning/logistics and
survey implementation. These include datasets on population, lists of
sampling clusters, map datasets for spatial sampling, and previous survey
datasets that can be used for estimating indicator variance and design
effects. This package contains relevant datasets for use in designing
surveys in Kenya.
}
