% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lhz}
\alias{lhz}
\title{lhz}
\format{A SpatialPolygonsDataFrame with 24 features:
\describe{
\item{\code{FNID}}{Feature identifier}
\item{\code{EFF_YEAR}}{Year}
\item{\code{COUNTRY}}{Name of country}
\item{\code{LZNUM}}{Livelihood zone number}
\item{\code{LZCODE}}{Livelihood zone code}
\item{\code{LZNAMEEN}}{Livelihood zone name - English}
\item{\code{LZNAMEFR}}{Livelihood zone name - French}
\item{\code{LZNAMESP}}{Livelihood zone name - Spanish}
\item{\code{LZNAMEPT}}{Livelihood zone name - Portuguese}
\item{\code{CLASS}}{Livelihood zone class}
}}
\source{
\url{http://fews.net/east-africa/kenya/livelihood-zone-map/march-2011}
}
\usage{
lhz
}
\description{
Livelihood zones map of Kenya
}
\examples{
# View lhz SpatialPolygonsDataFrame
lhz

}
\keyword{datasets}
