% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlements}
\alias{settlements}
\title{settlements}
\format{A SpatialPointsDataFrame with 2675 features:
\describe{
\item{\code{Z__}}{Z__}
\item{\code{UFI}}{Unique feature identifier (UTM)}
\item{\code{UNI}}{Unique name identifier (UTM)}
\item{\code{DLAT}}{Latitude coordinates (degrees)}
\item{\code{DLONG}}{Longitude coordinates (degrees)}
\item{\code{LAT}}{Latitude coordinate (UTM)}
\item{\code{LONG}}{Longitude coordinate (UTM)}
\item{\code{MGRS}}{Military Grid Reference System code}
\item{\code{FC}}{Feature code}
\item{\code{DSG}}{Description code}
\item{\code{PC}}{Place code}
\item{\code{CC1}}{Country code 1}
\item{\code{ADM1}}{Administrative level 1 code}
\item{\code{ADM2}}{Administrative level 2 code}
\item{\code{DIM}}{DIM}
\item{\code{CC2}}{Country code 2}
\item{\code{NT}}{NT}
\item{\code{LC}}{LC}
\item{\code{SHORT_FORM}}{Short form}
\item{\code{GENERIC}}{Generic}
\item{\code{SORT_NAME}}{Sort name}
\item{\code{FULL_NAME}}{Full name}
\item{\code{FULL_NAME_}}{Full name}
\item{\code{NEWDLAT}}{New latitude (degrees)}
\item{\code{NEWDLONG}}{New longitude (degrees)}
\item{\code{CODE}}{Code}
\item{\code{DISTRICT}}{District name}
\item{\code{REGION}}{Region name}
\item{\code{LOCATION}}{Location name}
\item{\code{SUB_LOCATI}}{Sub-location name}
}}
\source{
\url{https://data.humdata.org/dataset/kenya-settlements-0}
}
\usage{
settlements
}
\description{
Settlements of Kenya
}
\keyword{datasets}
