% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ward}
\alias{ward}
\title{Administrative Level 3 (wards) map of Kenya}
\format{
A SpatialPolygonsDataFrame with 1450 features:
\describe{
\item{\code{gid}}{Geographic identifier}
\item{\code{pop2009}}{Ward population in 2009}
\item{\code{county}}{Name of county}
\item{\code{subcounty}}{Name of subcounty}
\item{\code{ward}}{Name of ward}
\item{\code{uid}}{Unique identifier}
\item{\code{scuid}}{Subcounty unique identifier}
\item{\code{cuid}}{County unique identifier}
}
}
\source{
\url{https://data.humdata.org/dataset/administrative-wards-in-kenya-1450}
}
\usage{
ward
}
\description{
Administrative Level 3 (wards) map of Kenya
}
\examples{
# View ward SpatialPolygonsDataFrame
ward

}
\keyword{datasets}
